requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
django-crontab==0.7.1
django-debug-toolbar==4.2.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
psycopg2==2.9.7
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
